refactor: Extract JWT and Redis code to separate jwt.py module
- Created new backend/jwt.py module containing:
  * JWT configuration and initialization
  * Redis connection for token blacklisting
  * JWT error handlers (expired, invalid, unauthorized)
  * Token blacklisting utility functions
  * Clean separation of JWT concerns

- Refactored backend/app.py:
  * Removed JWT configuration code
  * Removed Redis connection setup
  * Removed JWT error handlers
  * Updated imports to use jwt.py module
  * Updated logout route to use blacklist_token function
  * Cleaner, more focused application logic

- Benefits of refactoring:
  * Better separation of concerns
  * JWT functionality centralized in one module
  * Easier JWT configuration management
  * Redis connection isolated to JWT module
  * Cleaner app.py focused on routes and business logic
  * Better testability and maintainability

This refactoring improves code organization while maintaining
all existing JWT functionality and API endpoints.
